fix: corrected error in fixed
